# src/web/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Etu health monitoring",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order/new": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "List new orders",
                "description": "Method for get list new orders without doctor",
                "operationId": "app\\modules\\api\\controllers\\OrderController::actionNew",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderShortViewDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/current": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "List current orders (for doctors)",
                "description": "Method for get list current orders (by doctor)",
                "operationId": "app\\modules\\api\\controllers\\OrderController::actionCurrent",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderShortViewDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/discharged": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "List discharged orders (for doctors)",
                "description": "Method for get list discharged orders (by doctor)",
                "operationId": "app\\modules\\api\\controllers\\OrderController::actionDischarged",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderShortViewDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/create": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create order",
                "description": "Method for create order of treatment",
                "operationId": "app\\modules\\api\\controllers\\OrderController::actionCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreateForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "temperature": {
                                            "type": "float",
                                            "example": "37.7"
                                        },
                                        "symptoms": {
                                            "type": "text",
                                            "example": "Болит голова, живот, нога, глаза"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/accept": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Accept order by doctor",
                "description": "Method for accept order without doctor",
                "operationId": "app\\modules\\api\\controllers\\OrderController::actionAccept",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderAcceptForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "order_id": {
                                            "type": "int",
                                            "example": "5"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/task/all": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "List all task ",
                "description": "Method for get list all task without pagination",
                "operationId": "app\\modules\\api\\controllers\\TaskController::actionAll",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "100500"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "active": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskShortViewDto"
                                            }
                                        },
                                        "done": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskShortViewDto"
                                            }
                                        },
                                        "removed": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskShortViewDto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/task/create": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Create task by order",
                "description": "Method for create task for patient by order",
                "operationId": "app\\modules\\api\\controllers\\TaskController::actionCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TaskCreateForm"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskShortViewDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrderShortViewDto": {
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "example": "100500"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Василий",
                        "nullable": true
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Пупкин",
                        "nullable": true
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "Петрович",
                        "nullable": true
                    },
                    "temperature": {
                        "type": "number",
                        "format": "float",
                        "example": "36.6",
                        "nullable": true
                    },
                    "symptoms": {
                        "type": "string",
                        "example": "Болит голова, живот, нога, глаза"
                    },
                    "created_at": {
                        "type": "integer",
                        "example": "100500",
                        "nullable": true
                    },
                    "age": {
                        "type": "integer",
                        "example": "25",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "TaskShortViewDto": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "100500"
                    },
                    "order_id": {
                        "type": "integer",
                        "example": "100500"
                    },
                    "title": {
                        "type": "string",
                        "example": "Доктор Мом",
                        "nullable": true
                    },
                    "planned_at": {
                        "type": "string",
                        "example": "2021-05-25",
                        "nullable": true
                    },
                    "done_at": {
                        "type": "integer",
                        "example": "2021-05-25",
                        "nullable": true
                    },
                    "removed_at": {
                        "type": "integer",
                        "example": "2021-05-25",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OrderAcceptForm": {
                "required": [
                    "order_id"
                ],
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "example": "100500"
                    }
                },
                "type": "object"
            },
            "OrderCreateForm": {
                "required": [
                    "temperature",
                    "symptoms"
                ],
                "properties": {
                    "temperature": {
                        "type": "number",
                        "format": "float",
                        "example": "37.6"
                    },
                    "symptoms": {
                        "type": "string",
                        "example": "Болит голова, живот, нога, глаза"
                    }
                },
                "type": "object"
            },
            "TaskCreateForm": {
                "required": [
                    "order_id",
                    "date_to",
                    "title",
                    "quantity"
                ],
                "properties": {
                    "order_id": {
                        "type": "integer",
                        "example": "100500"
                    },
                    "date_to": {
                        "type": "integer",
                        "example": "2021-05-25"
                    },
                    "title": {
                        "type": "string",
                        "example": "Арбидол"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": "Количество"
                    }
                },
                "type": "object"
            },
...
        }
    }
}